refactor(#21): Template 내 하위 컴포넌트 분리
