+++ src/components/templates/BookReivew/Article.tsx
@@ -1,0 +1,19 @@
+import { ReactNode } from 'react';
+import * as s from './style';
+
+export interface BookReviewArticleProps {
+  sejulViewer: ReactNode;
+  contentViewer: ReactNode;
+}
+
+const BookReviewArticle = ({
+  sejulViewer,
+  contentViewer,
+}: BookReviewArticleProps) => (
+  <s.Article>
+    {sejulViewer}
+    {contentViewer}
+  </s.Article>
+);
+
+export default BookReviewArticle;
